registra huella de carbono

+++ app/src/main/java/com/bancomer/bbva/bbvamovilidad/data/api/request/Detalle.kt
@@ -4,7 +4,7 @@ data class Detalle(
     var fhFinRecorrido: Long? = null,
     var fhIniRecorrido: Long? = null,
     var idMedioTraslado: Int? = null,
-    var kmRecorrido: Int? = null,
+    var kmRecorrido: Float? = null,
     var origenLatitud: Double? = null,
     var origenLongitud: Double? = null,
     var paradaLatitud: Double? = null,
